refactor(benthosmetrics): extract parser, types, and windowing from benthos_monitor
Break the import cycle by moving the metrics parser (ParseMetricsFromBytes, TailInt, extractLabel), its value types (Metrics, HealthCheck, BenthosMetrics, ...), and the throughput windowing (BenthosMetricsState) into a new pkg/service/benthosmetrics package. benthos_monitor keeps type aliases for the moved types so all existing importers compile unchanged. The previously unexported connStatus/readyResponse/versionResponse are exported as ConnStatus/ReadyResponse/VersionResponse (Go cannot alias unexported foreign identifiers), with unexported aliases left behind for internal namers (e.g. mock.go's []connStatus).

Adds the net-new Scan{Metrics, HealthCheck, MetricsAvailable} type that the FSMv2 benthos_monitor worker will publish; not populated yet (Observe comes in a later commit).

No behavior change. Existing benthos_monitor + metrics_state suites pass; benthosmetrics imports nothing from benthos_monitor.
