+++ umh-core/pkg/service/benthos_monitor/benthos_monitor.go
@@ -39,197 +39,13 @@ import (
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/standarderrors"
 
 	s6fsm "github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsm/s6"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/service/benthosmetrics"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/service/filesystem"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/service/monitor"
 	s6service "github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/service/s6"
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 )
-
-// Metrics contains information about the metrics of the Benthos service.
-// Inputs and Outputs are keyed by benthos `path` label (one entry per leaf
-// AsyncReader / AsyncWriter). For switch/broker/fallback configs, multiple
-// entries exist; for a single-input or single-output config there is one
-// entry at "root.input" / "root.output". Aggregate values are exposed via
-// the *Total helpers on Metrics.
-type Metrics struct {
-	Process ProcessMetrics            `json:"process,omitempty"`
-	Outputs map[string]OutputInstance `json:"outputs,omitempty"`
-	Inputs  map[string]InputInstance  `json:"inputs,omitempty"`
-}
-
-// InputInstance contains per-path input metrics for one benthos AsyncReader.
-type InputInstance struct {
-	Path             string  `json:"path"`
-	Label            string  `json:"label"`
-	ConnectionFailed int64   `json:"connection_failed"`
-	ConnectionLost   int64   `json:"connection_lost"`
-	ConnectionUp     int64   `json:"connection_up"`
-	LatencyNS        Latency `json:"latency_ns"`
-	Received         int64   `json:"received"`
-}
-
-// OutputInstance contains per-path output metrics for one benthos AsyncWriter.
-type OutputInstance struct {
-	Path             string  `json:"path"`
-	Label            string  `json:"label"`
-	BatchSent        int64   `json:"batch_sent"`
-	ConnectionFailed int64   `json:"connection_failed"`
-	ConnectionLost   int64   `json:"connection_lost"`
-	ConnectionUp     int64   `json:"connection_up"`
-	Error            int64   `json:"error"`
-	LatencyNS        Latency `json:"latency_ns"`
-	Sent             int64   `json:"sent"`
-}
-
-// ProcessMetrics contains processor-specific metrics.
-type ProcessMetrics struct {
-	Processors map[string]ProcessorMetrics `json:"processors"` // key is the processor path (e.g. "root.pipeline.processors.0")
-}
-
-// ProcessorMetrics contains metrics for a single processor.
-type ProcessorMetrics struct {
-	Label         string  `json:"label"`
-	Received      int64   `json:"received"`
-	BatchReceived int64   `json:"batch_received"`
-	Sent          int64   `json:"sent"`
-	BatchSent     int64   `json:"batch_sent"`
-	Error         int64   `json:"error"`
-	LatencyNS     Latency `json:"latency_ns"`
-}
-
-// Latency contains latency metrics.
-type Latency struct {
-	P50   float64 `json:"p50"`   // 50th percentile
-	P90   float64 `json:"p90"`   // 90th percentile
-	P99   float64 `json:"p99"`   // 99th percentile
-	Sum   float64 `json:"sum"`   // Total sum
-	Count int64   `json:"count"` // Number of samples
-}
-
-// InputReceivedTotal returns the sum of Received across every input instance.
-func (m Metrics) InputReceivedTotal() int64 {
-	var total int64
-	for _, in := range m.Inputs {
-		total += in.Received
-	}
-
-	return total
-}
-
-// InputConnectionUpTotal returns the sum of ConnectionUp across every input.
-func (m Metrics) InputConnectionUpTotal() int64 {
-	var total int64
-	for _, in := range m.Inputs {
-		total += in.ConnectionUp
-	}
-
-	return total
-}
-
-// InputConnectionLostTotal returns the sum of ConnectionLost across every input.
-func (m Metrics) InputConnectionLostTotal() int64 {
-	var total int64
-	for _, in := range m.Inputs {
-		total += in.ConnectionLost
-	}
-
-	return total
-}
-
-// OutputSentTotal returns the sum of Sent across every output instance.
-func (m Metrics) OutputSentTotal() int64 {
-	var total int64
-	for _, out := range m.Outputs {
-		total += out.Sent
-	}
-
-	return total
-}
-
-// OutputBatchSentTotal returns the sum of BatchSent across every output.
-func (m Metrics) OutputBatchSentTotal() int64 {
-	var total int64
-	for _, out := range m.Outputs {
-		total += out.BatchSent
-	}
-
-	return total
-}
-
-// OutputConnectionUpTotal returns the sum of ConnectionUp across every output.
-func (m Metrics) OutputConnectionUpTotal() int64 {
-	var total int64
-	for _, out := range m.Outputs {
-		total += out.ConnectionUp
-	}
-
-	return total
-}
-
-// OutputConnectionLostTotal returns the sum of ConnectionLost across every output.
-func (m Metrics) OutputConnectionLostTotal() int64 {
-	var total int64
-	for _, out := range m.Outputs {
-		total += out.ConnectionLost
-	}
-
-	return total
-}
-
-// OutputErrorTotal returns the sum of Error across every output. Used by
-// benthos.IsMetricsErrorFree to decide whether the FSM should treat the
-// component as healthy.
-func (m Metrics) OutputErrorTotal() int64 {
-	var total int64
-	for _, out := range m.Outputs {
-		total += out.Error
-	}
-
-	return total
-}
-
-// HealthCheck contains information about the health of the Benthos service
-// https://docs.redpanda.com/redpanda-connect/guides/monitoring/
-type HealthCheck struct {
-	// Version contains the version of the Benthos service
-	Version string
-	// ReadyError contains any error message from the ready check
-	ReadyError string `json:"ready_error,omitempty"`
-	// ConnectionStatuses contains the detailed connection status of inputs and outputs
-	ConnectionStatuses []connStatus `json:"connection_statuses,omitempty"`
-	// IsLive is true if the Benthos service is live
-	IsLive bool
-	// IsReady is true if the Benthos service is ready to process data
-	IsReady bool
-}
-
-// versionResponse represents the JSON structure returned by the /version endpoint.
-type versionResponse struct {
-	Version string `json:"version"`
-	Built   string `json:"built"`
-}
-
-// readyResponse represents the JSON structure returned by the /ready endpoint.
-type readyResponse struct {
-	Error    string       `json:"error,omitempty"`
-	Statuses []connStatus `json:"statuses"`
-}
-
-type connStatus struct {
-	Label     string `json:"label"`
-	Path      string `json:"path"`
-	Error     string `json:"error,omitempty"`
-	Connected bool   `json:"connected"`
-}
-
-// BenthosMetrics contains information about the metrics of the Benthos service.
-type BenthosMetrics struct {
-	// MetricsState contains the state of the metrics
-	MetricsState *BenthosMetricsState
-	// Metrics contains the metrics of the Benthos service
-	Metrics Metrics
-}
 
 // ServiceInfo contains information about a benthos service.
 type ServiceInfo struct {
@@ -356,7 +172,7 @@ func NewBenthosMonitorService(benthosName string, opts ...BenthosMonitorServiceO
 
 	service := &BenthosMonitorService{
 		logger:       logger.For(managerName),
-		metricsState: NewBenthosMetricsState(),
+		metricsState: benthosmetrics.NewBenthosMetricsState(),
 		s6Manager:    s6fsm.NewS6Manager(logger.ComponentBenthosMonitorService),
 		s6Service:    s6service.NewDefaultService(),
 		benthosName:  benthosName,
@@ -1002,416 +818,13 @@ func ParseMetricsData(dataReader io.Reader) (Metrics, error) {
 		return Metrics{}, curlError
 	}
 
-	metrics, err := ParseMetricsFromBytes(data)
+	metrics, err := benthosmetrics.ParseMetricsFromBytes(data)
 	if err != nil {
 		return Metrics{}, fmt.Errorf("failed to parse metrics data: %w", err)
 	}
 
 	return metrics, nil
 }
-
-// --- helpers ---------------------------------------------------------------
-
-// TailInt returns the integer found **after the final space** in a Prometheus
-// text-formatted line.
-//
-// Benthos' metric stream is deliberately terse; for counters and gauges the
-// value we need is always the token after the last space.  A micro-parser that
-// walks backwards from the end lets us avoid a full `strings.Fields` split
-// (~3× faster and zero allocations on the hot-path).
-func TailInt(line []byte) (int64, error) {
-	i := bytes.LastIndexByte(line, ' ')
-	if i == -1 {
-		return 0, errors.New("failed to find space in line")
-	}
-
-	s := string(line[i+1:])
-	if strings.ContainsAny(s, "eE") {
-		f, err := strconv.ParseFloat(s, 64)
-		if err != nil {
-			return 0, fmt.Errorf("failed to parse float: %w", err)
-		}
-
-		return int64(f), nil
-	}
-
-	var v int64
-
-	for _, c := range s {
-		if c < '0' || c > '9' {
-			break
-		}
-
-		v = v*10 + int64(c-'0')
-	}
-
-	return v, nil
-}
-
-// ---------------------------------------------------------------------------
-
-// ParseMetricsFromBytes parses the prometheus text-format payload benthos
-// returns and produces a per-path Metrics. For switch/broker/fallback
-// outputs the input emits N distinct series per metric, one per leaf
-// path; the parser stores each in its own InputInstance / OutputInstance
-// entry.
-func ParseMetricsFromBytes(raw []byte) (Metrics, error) {
-	var m Metrics
-
-	m.Process.Processors = make(map[string]ProcessorMetrics, 8)
-	m.Inputs = make(map[string]InputInstance, 1)
-	m.Outputs = make(map[string]OutputInstance, 1)
-
-	lineStart := 0
-
-	for i, b := range raw {
-		if b != '\n' && i != len(raw)-1 {
-			continue
-		}
-
-		end := i
-		if b != '\n' { // EOF w/o \n
-			end++
-		}
-
-		line := raw[lineStart:end]
-		lineStart = i + 1
-
-		if len(line) == 0 || line[0] == '#' {
-			continue
-		}
-
-		nameEnd := bytes.IndexByte(line, '{')
-		if nameEnd == -1 {
-			nameEnd = bytes.IndexByte(line, ' ')
-		}
-
-		if nameEnd == -1 {
-			continue
-		}
-
-		name := string(line[:nameEnd])
-
-		// All input_* and output_* lines carry a `path` label.
-		labelsRegion := line[nameEnd:]
-
-		switch {
-		case strings.HasPrefix(name, "input_"):
-			path := extractLabel(labelsRegion, "path")
-			if path == "" {
-				path = "root.input"
-			}
-
-			label := extractLabel(labelsRegion, "label")
-
-			in := m.Inputs[path]
-			if in.Path == "" {
-				in.Path = path
-			}
-
-			if in.Label == "" && label != "" {
-				in.Label = label
-			}
-
-			switch name {
-			case "input_connection_failed":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse input connection failed: %w", err)
-				}
-
-				in.ConnectionFailed = count
-			case "input_connection_lost":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse input connection lost: %w", err)
-				}
-
-				in.ConnectionLost = count
-			case "input_connection_up":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse input connection up: %w", err)
-				}
-
-				in.ConnectionUp = count
-			case "input_received":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse input received: %w", err)
-				}
-
-				in.Received = count
-			case "input_latency_ns":
-				switch extractLabel(labelsRegion, "quantile") {
-				case "0.5":
-					p50, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse input latency ns p50: %w", err)
-					}
-
-					in.LatencyNS.P50 = float64(p50)
-				case "0.9":
-					p90, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse input latency ns p90: %w", err)
-					}
-
-					in.LatencyNS.P90 = float64(p90)
-				case "0.99":
-					p99, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse input latency ns p99: %w", err)
-					}
-
-					in.LatencyNS.P99 = float64(p99)
-				}
-			case "input_latency_ns_sum":
-				sum, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse input latency ns sum: %w", err)
-				}
-
-				in.LatencyNS.Sum = float64(sum)
-			case "input_latency_ns_count":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse input latency ns count: %w", err)
-				}
-
-				in.LatencyNS.Count = count
-			}
-
-			m.Inputs[path] = in
-
-		case strings.HasPrefix(name, "output_"):
-			path := extractLabel(labelsRegion, "path")
-			if path == "" {
-				path = "root.output"
-			}
-
-			label := extractLabel(labelsRegion, "label")
-
-			out := m.Outputs[path]
-			if out.Path == "" {
-				out.Path = path
-			}
-
-			if out.Label == "" && label != "" {
-				out.Label = label
-			}
-
-			switch name {
-			case "output_batch_sent":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse output batch sent: %w", err)
-				}
-
-				out.BatchSent = count
-			case "output_connection_failed":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse output connection failed: %w", err)
-				}
-
-				out.ConnectionFailed = count
-			case "output_connection_lost":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse output connection lost: %w", err)
-				}
-
-				out.ConnectionLost = count
-			case "output_connection_up":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse output connection up: %w", err)
-				}
-
-				out.ConnectionUp = count
-			case "output_error":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse output error: %w", err)
-				}
-
-				out.Error = count
-			case "output_sent":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse output sent: %w", err)
-				}
-
-				out.Sent = count
-			case "output_latency_ns":
-				switch extractLabel(labelsRegion, "quantile") {
-				case "0.5":
-					p50, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse output latency ns p50: %w", err)
-					}
-
-					out.LatencyNS.P50 = float64(p50)
-				case "0.9":
-					p90, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse output latency ns p90: %w", err)
-					}
-
-					out.LatencyNS.P90 = float64(p90)
-				case "0.99":
-					p99, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse output latency ns p99: %w", err)
-					}
-
-					out.LatencyNS.P99 = float64(p99)
-				}
-			case "output_latency_ns_sum":
-				sum, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse output latency ns sum: %w", err)
-				}
-
-				out.LatencyNS.Sum = float64(sum)
-			case "output_latency_ns_count":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse output latency ns count: %w", err)
-				}
-
-				out.LatencyNS.Count = count
-			}
-
-			m.Outputs[path] = out
-
-		default:
-			if !bytes.HasPrefix(line, []byte("processor_")) {
-				continue
-			}
-
-			path := extractLabel(labelsRegion, "path")
-			if path == "" {
-				continue
-			}
-
-			pm := m.Process.Processors[path]
-			if pm.Label == "" {
-				pm.Label = extractLabel(labelsRegion, "label")
-			}
-
-			switch name {
-			case "processor_received":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse processor received: %w", err)
-				}
-
-				pm.Received = count
-			case "processor_batch_received":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse processor batch received: %w", err)
-				}
-
-				pm.BatchReceived = count
-			case "processor_sent":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse processor sent: %w", err)
-				}
-
-				pm.Sent = count
-			case "processor_batch_sent":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse processor batch sent: %w", err)
-				}
-
-				pm.BatchSent = count
-			case "processor_error":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse processor error: %w", err)
-				}
-
-				pm.Error = count
-			case "processor_latency_ns":
-				switch extractLabel(labelsRegion, "quantile") {
-				case "0.5":
-					p50, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse processor latency ns p50: %w", err)
-					}
-
-					pm.LatencyNS.P50 = float64(p50)
-				case "0.9":
-					p90, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse processor latency ns p90: %w", err)
-					}
-
-					pm.LatencyNS.P90 = float64(p90)
-				case "0.99":
-					p99, err := TailInt(line)
-					if err != nil {
-						return Metrics{}, fmt.Errorf("failed to parse processor latency ns p99: %w", err)
-					}
-
-					pm.LatencyNS.P99 = float64(p99)
-				}
-			case "processor_latency_ns_sum":
-				sum, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse processor latency ns sum: %w", err)
-				}
-
-				pm.LatencyNS.Sum = float64(sum)
-			case "processor_latency_ns_count":
-				count, err := TailInt(line)
-				if err != nil {
-					return Metrics{}, fmt.Errorf("failed to parse processor latency ns count: %w", err)
-				}
-
-				pm.LatencyNS.Count = count
-			}
-
-			m.Process.Processors[path] = pm
-		}
-	}
-
-	return m, nil
-}
-
-// extractLabel scans the raw label-set of a Prometheus series (the `{…}`
-// portion) and returns the *value* of a given label key.
-//
-// We read **only** the labels we care about (e.g. `path` or `quantile`), so
-// this hand-rolled matcher is an order of magnitude faster than allocating a
-// full `map[string]string` for every series.
-func extractLabel(b []byte, key string) string {
-	// expects {label="...",path="..."} order irrelevant
-	key += "=\""
-
-	i := bytes.Index(b, []byte(key))
-	if i == -1 {
-		return ""
-	}
-
-	i += len(key)
-
-	j := bytes.IndexByte(b[i:], '"')
-	if j == -1 {
-		return ""
-	}
-	// return unsafeString(b[i : i+j])
-	return string(b[i : i+j])
-}
-
-// unsafeString converts a []byte to string without allocation.
-// Use only for read-only parsing.
-// func unsafeString(b []byte) string { return *(*string)(unsafe.Pointer(&b)) }
 
 // Status checks the status of a benthos service.
 func (s *BenthosMonitorService) Status(ctx context.Context, services serviceregistry.Provider, tick uint64) (ServiceInfo, error) {
@@ -1590,7 +1003,7 @@ func (s *BenthosMonitorService) RemoveBenthosMonitorFromS6Manager(ctx context.Co
 	}
 
 	// Clean up the metrics state
-	s.metricsState = NewBenthosMetricsState()
+	s.metricsState = benthosmetrics.NewBenthosMetricsState()
 
 	return nil
 }
