+++ umh-core/pkg/service/benthosmetrics/metrics_state.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package benthos_monitor
+package benthosmetrics
 
 // ComponentThroughput tracks throughput metrics for a single component.
 type ComponentThroughput struct {
